Refacotr completeTask function
