+++ src/features/components/TaskProgress/TaskCard.tsx
@@ -1,8 +1,9 @@
 import React from 'react'
 import type { Task, CSSProperties } from '../../../types'
 import {TASK_PROGRESS_ID} from '../../../constants/app'
-import { useRecoilState } from 'recoil'  // Ditambahkan
-import { tasksState } from '../../TaskAtoms'  // Ditambahkan
+//import { useRecoilState } from 'recoil'  // Ditambahkan
+//import { tasksState } from '../../TaskAtoms'  // Ditambahkan
+import {useTasksAction} from '../../hooks/Tasks'
 
 
 interface TaskCardProps {
@@ -35,15 +36,16 @@ const getArrowPositionStyle = (progressOrder: number): React.CSSProperties => {
 }
 
 const TaskCard = ({ task }: TaskCardProps): JSX.Element => {
-  const [tasks, setTasks] = useRecoilState<Task[]>(tasksState)
-  const completeTask = (taskId: number): void => {
-    const updatedTasks: Task[] = tasks.map((task) =>
-      task.id === taskId
-        ? { ...task, progressOrder: TASK_PROGRESS_ID.COMPLETED }
-        : task,
-    )
-    setTasks(updatedTasks)
-  }
+  // const [tasks, setTasks] = useRecoilState<Task[]>(tasksState)
+  // const completeTask = (taskId: number): void => {
+  //   const updatedTasks: Task[] = tasks.map((task) =>
+  //     task.id === taskId
+  //       ? { ...task, progressOrder: TASK_PROGRESS_ID.COMPLETED }
+  //       : task,
+  //   )
+  //   setTasks(updatedTasks)
+  // }
+  const {completeTask} = useTasksAction()
 
   return (
     <div style={styles.taskCard}>
